+++ test/builders/process.ts
@@ -17,8 +17,8 @@ export const DEFAULT_PROCESS_MODE = ProcessMode.make()
 export const DEFAULT_ENVELOPE_TYPE = ProcessEnvelopeType.make()
 export const DEFAULT_CENSUS_ORIGIN = ProcessCensusOrigin.OFF_CHAIN
 export const DEFAULT_METADATA_CONTENT_HASHED_URI = "ipfs://1234,https://server/uri!0987654321"
-export const DEFAULT_MERKLE_ROOT = "0x123456789"
-export const DEFAULT_MERKLE_TREE_CONTENT_HASHED_URI = "ipfs://1234,https://server/uri!1234567812345678"
+export const DEFAULT_CENSUS_ROOT = "0x123456789"
+export const DEFAULT_CENSUS_TREE_CONTENT_HASHED_URI = "ipfs://1234,https://server/uri!1234567812345678"
 export const DEFAULT_START_BLOCK = 12341234
 export const DEFAULT_BLOCK_COUNT = 500000
 export const DEFAULT_QUESTION_COUNT = 5
@@ -40,8 +40,8 @@ export default class ProcessBuilder {
     predecessorInstanceAddress: string = DEFAULT_PREDECESSOR_INSTANCE_ADDRESS
     enabled: boolean = true
     metadata: string = DEFAULT_METADATA_CONTENT_HASHED_URI
-    merkleRoot: string = DEFAULT_MERKLE_ROOT
-    merkleTree: string = DEFAULT_MERKLE_TREE_CONTENT_HASHED_URI
+    censusRoot: string = DEFAULT_CENSUS_ROOT
+    censusTree: string = DEFAULT_CENSUS_TREE_CONTENT_HASHED_URI
     startBlock: number = DEFAULT_START_BLOCK
     blockCount: number = DEFAULT_BLOCK_COUNT
     mode: IProcessMode = DEFAULT_PROCESS_MODE
@@ -111,8 +111,8 @@ export default class ProcessBuilder {
                 envelopeType: this.envelopeType,
                 censusOrigin: this.censusOrigin,
                 metadata: this.metadata,
-                censusMerkleRoot: this.merkleRoot,
-                censusMerkleTree: this.merkleTree,
+                censusRoot: this.censusRoot,
+                censusUri: this.censusTree,
                 startBlock: this.startBlock,
                 blockCount: this.blockCount,
                 questionCount: this.questionCount,
@@ -226,8 +226,8 @@ export default class ProcessBuilder {
             envelopeType: DEFAULT_ENVELOPE_TYPE,
             censusOrigin: DEFAULT_CENSUS_ORIGIN,
             metadata: DEFAULT_METADATA_CONTENT_HASHED_URI,
-            censusMerkleRoot: DEFAULT_MERKLE_ROOT,
-            censusMerkleTree: DEFAULT_MERKLE_TREE_CONTENT_HASHED_URI,
+            censusRoot: DEFAULT_CENSUS_ROOT,
+            censusUri: DEFAULT_CENSUS_TREE_CONTENT_HASHED_URI,
             startBlock: DEFAULT_START_BLOCK,
             blockCount: DEFAULT_BLOCK_COUNT,
             questionCount: DEFAULT_QUESTION_COUNT,